requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.3
pytest==8.3.3
httpx==0.27.2
celery==5.4.0
redis==5.0.8
scikit-learn==1.5.2
python-dotenv==1.0.1
matplotlib==3.9.2
numpy==1.26.4